Cleanup: improve integer types
Reviewers: mont29, brecht

Differential Revision: https://developer.blender.org/D9135

+++ source/blender/blenkernel/BKE_idtype.h
@@ -114,14 +114,14 @@ typedef struct IDTypeInfo {
    * Bitflag matching id_code, used for filtering (e.g. in file browser), see DNA_ID.h's
    * FILTER_ID_XX enums.
    */
-  int64_t id_filter;
+  uint64_t id_filter;
 
   /**
    * Define the position of this data-block type in the virtual list of all data in a Main that is
    * returned by `set_listbasepointers()`.
    * Very important, this has to be unique and below INDEX_ID_MAX, see DNA_ID.h.
    */
-  short main_listbase_index;
+  int main_listbase_index;
 
   /** Memory size of a data-block of that type. */
   size_t struct_size;
@@ -134,7 +134,7 @@ typedef struct IDTypeInfo {
   const char *translation_context;
 
   /** Generic info flags about that data-block type. */
-  int flags;
+  uint32_t flags;
 
   /* ********** ID management callbacks ********** */
